requirements.txt
aiohttp==3.7.4.post0
ujson==4.2.0
requests==2.26.0
aioboto3==9.2.2
aiofiles==0.7.0
pyfailsafe==0.6.0
pytz==2021.3